Refactor to better separate quex types
