+++ src/routes/mod.rs
@@ -1,0 +1,5 @@
+mod questionnaires;
+pub(crate) mod index;
+pub(crate) mod net_promoter_score;
+pub(crate) mod attrakdiff;
+pub(crate) mod system_usability_score;
